+++ backend/src/middleware/api/database.js
@@ -1,8 +1,8 @@
 import mongoose from "mongoose";
 import Brand from "../../database/models/Brand.js";
 import Product from "../../database/models/Product.js";
+import Sale from "../../database/models/Sale.js";
 import Variant from "../../database/models/Variant.js";
-import Sale from "../../database/models/Sale.js"
 
 /////////////////////
 // CREATING MODELS //
@@ -17,16 +17,16 @@ import Sale from "../../database/models/Sale.js"
  * @returns {Brand} The brand object created.
  */
 export async function createBrand(options) {
-    const brand = new Brand({
-        name: options.name,
-        category: options.category,
-    });
+  const brand = new Brand({
+    name: options.name,
+    category: options.category,
+  });
 
-    await brand.save();
-    console.log("Created brand:");
-    console.log(JSON.stringify(brand));
+  await brand.save();
+  console.log("Created brand:");
+  console.log(JSON.stringify(brand));
 
-    return brand;
+  return brand;
 }
 
 /**
@@ -43,37 +43,37 @@ export async function createBrand(options) {
  * @returns {Product} The product object created.
  */
 export async function createProduct(options) {
-    const brands = await Brand.find({
-        name: options.brandName,
-        category: options.brandCategory,
-    });
+  const brands = await Brand.find({
+    name: options.brandName,
+    category: options.brandCategory,
+  });
 
-    if (brands.length === 0)
-        throw new Error("Brand is not found.");
+  if (brands.length === 0)
+    throw new Error("Brand is not found.");
 
-    const variants = await Variant.find({
-        name: options.variantName,
-    });
+  const variants = await Variant.find({
+    name: options.variantName,
+  });
 
-    if (variants.length === 0)
-        throw new Error("Variant is not found.");
+  if (variants.length === 0)
+    throw new Error("Variant is not found.");
 
-    const brand = brands[0];
-    const variant = variants[0];
-    const product = new Product({
-        brandId: brand._id,
-        variantId: variant._id,
-        name: options.name,
-        price: options.price,
-        quantity: options.quantity,
-        expiration: options.expiration,
-    });
+  const brand = brands[0];
+  const variant = variants[0];
+  const product = new Product({
+    brandId: brand._id,
+    variantId: variant._id,
+    name: options.name,
+    price: options.price,
+    quantity: options.quantity,
+    expiration: options.expiration,
+  });
 
-    await product.save();
-    console.log("Created product:");
-    console.log(JSON.stringify(product));
+  await product.save();
+  console.log("Created product:");
+  console.log(JSON.stringify(product));
 
-    return product;
+  return product;
 }
 
 /**
@@ -86,17 +86,17 @@ export async function createProduct(options) {
  * @returns {Variant} The variant object created.
  */
 export async function createVariant(options) {
-    const variant = new Variant({
-        name: options.name,
-        description: options.description,
-        category: options.category,
-    });
+  const variant = new Variant({
+    name: options.name,
+    description: options.description,
+    category: options.category,
+  });
 
-    await variant.save();
-    console.log("Created variant:");
-    console.log(JSON.stringify(variant));
+  await variant.save();
+  console.log("Created variant:");
+  console.log(JSON.stringify(variant));
 
-    return variant;
+  return variant;
 }
 
 /**
@@ -106,12 +106,12 @@ export async function createVariant(options) {
  * @returns {Array<Object>} List of brand documents matching the filter.
  */
 export async function readBrands(filter = {}) {
-    // Use the Brand model to query the database
-    const brands = await Brand.find(filter);
-    // console.log("Fetched brands:");
-    // console.log(JSON.stringify(brands, null, 2));
+  // Use the Brand model to query the database
+  const brands = await Brand.find(filter);
+  // console.log("Fetched brands:");
+  // console.log(JSON.stringify(brands, null, 2));
 
-    return brands;
+  return brands;
 }
 
 /**
@@ -122,13 +122,13 @@ export async function readBrands(filter = {}) {
  * @returns {Array<Object>} List of product documents matching the filter.
  */
 export async function readProducts(filter = {}) {
-    const products = await Product.find(filter);
-    return products;
+  const products = await Product.find(filter);
+  return products;
 }
 
 export async function readSales(filter = {}) {
-    const sales = await Sale.find(filter);
-    return sales;
+  const sales = await Sale.find(filter);
+  return sales;
 }
 
 /**
@@ -137,9 +137,9 @@ export async function readSales(filter = {}) {
  */
 export async function deleteProductById(id) {
 
-    await Product.findByIdAndDelete(id)
+  await Product.findByIdAndDelete(id)
 
-    console.log(`Product with ID ${JSON.stringify(id)} successfully deleted.`)
+  console.log(`Product with ID ${JSON.stringify(id)} successfully deleted.`)
 }
 
 /**
@@ -148,9 +148,9 @@ export async function deleteProductById(id) {
  */
 export async function deleteSalesById(id) {
 
-    await Sale.findByIdAndDelete(id)
+  await Sale.findByIdAndDelete(id)
 
-    console.log(`Product with ID ${JSON.stringify(id)} successfully deleted.`)
+  console.log(`Product with ID ${JSON.stringify(id)} successfully deleted.`)
 }
 
 
@@ -161,10 +161,10 @@ export async function deleteSalesById(id) {
  * @param {String}
  */
 export async function editProductById(filter) {
-    const { id, ...filterWithoutId } = filter;
-    await Product.findOneAndUpdate({ _id: filter.id },
-        { $set: { ...filterWithoutId } }, // Update: Change the email
-        { returnDocument: "after" })
+  const { id, ...filterWithoutId } = filter;
+  await Product.findOneAndUpdate({ _id: filter.id },
+    { $set: { ...filterWithoutId } }, // Update: Change the email
+    { returnDocument: "after" })
 }
 
 /**
@@ -175,30 +175,30 @@ export async function editProductById(filter) {
  * @param {String} filter.productId.
  */
 export async function createSale(filter) {
-    const brand = await Brand.findOne({ _id: filter.brandId });
-    const product = await Product.findOne({ _id: filter.productId });
+  const brand = await Brand.findOne({ _id: filter.brandId });
+  const product = await Product.findOne({ _id: filter.productId });
 
-    const totalSale = filter.sale * product["price"]
+  const totalSale = filter.sale * product["price"]
 
-    const sale = new Sale({
-        brandId: filter.brandId,
-        productId: filter.productId,
-        name: product["name"],
-        category: brand["category"],
-        quantity: filter.sale,
-        price: product["price"],
-        total: totalSale,
-        date: new Date(),
-    });
+  const sale = new Sale({
+    brandId: filter.brandId,
+    productId: filter.productId,
+    name: product["name"],
+    category: brand["category"],
+    quantity: filter.sale,
+    price: product["price"],
+    total: totalSale,
+    date: new Date(),
+  });
 
-    await sale.save()
+  await sale.save()
 
-    console.log("brand", brand)
-    console.log("product", product)
+  console.log("brand", brand)
+  console.log("product", product)
 
-    // await sale.save();
-    console.log("Created sale:");
-    console.log(JSON.stringify(sale));
+  // await sale.save();
+  console.log("Created sale:");
+  console.log(JSON.stringify(sale));
 
-    return sale;
+  return sale;
 }
